Remove erroneous sub-command execution
Node’s [`spawn`][1] is intended for running external shell executables
in sub-processes, not for executing node functions, as seemed the
intention in this code.

This change removes the use of spawn, instead calling `build` directly.
This change should resolve #18.

This could be adapted to use [`fork`][2] instead (which is designed to
execute a function in a separate v8 process), but  there doesn’t seem
to be good reason for that here.

[1]: https://nodejs.org/dist/v8.2.1/docs/api/child_process.html#child_process_child_process_spawn_command_args_options
[2]: https://nodejs.org/dist/latest-v8.x/docs/api/child_process.html#child_process_child_process_fork_modulepath_args_options